swap_nodes_in_pairs/swap_nodes_in_pairs_test.go
// https://leetcode.com/problems/swap-nodes-in-pairs/

package swap_nodes_in_pairs

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestSwapPairs(t *testing.T) {
	testCases := []struct {
		name     string
		head     *ListNode
		expected *ListNode
	}{
		{
			name: "Test Swap Pairs #1",
			head: &ListNode{
				Val: 1,
				Next: &ListNode{
					Val: 2,
					Next: &ListNode{
						Val: 3,
						Next: &ListNode{
							Val: 4,
						},
					},
				},
			},
			expected: &ListNode{
				Val: 2,
				Next: &ListNode{
					Val: 1,
					Next: &ListNode{
						Val: 4,
						Next: &ListNode{
							Val: 3,
						},
					},
				},
			},
		},
	}

	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			// assert.Equal(t, tc.expected, swapPairs(tc.head))
			assert.Equal(t, tc.expected, swapPairs(tc.head))
		})
	}
}
